Reorged files a bit.
